devtools/client - improve organization
